Refactor wallet_tests.cpp
Change `testTx()` to `broadcastAndProcessBlock()` because this funciton is copied from other test file and the usage in wallet_tests.cpp is different from the purpose of it.
Here, the function is just used for adding a tx into mempool and create new block. So this commit changes the function for the actual useage.