Add a CreateTransaction test
Add CScriptVisitor constructor for non-colored coin

Before this change, we could use common interface for non-colored coin with colored coin, but it was not explicit that we can use it for non-colored coin.

Make CreateTransaction raise an error when a recipient that receives colored coin would be set as a target of subtracting fee

Add a function that issues non-reissunable colored coin to wallet test setup

Add a unit test for creating colored coin transaction

Add Colored coin support to CWallet::CreateTransaction()

Make it create change outputs for colored coins

Support change-position argument on CreateTransaction() for multiple change outputs with colored coin

Add test to create outputs which has small value like dust output in colored coin

Add lock cs_wallet when sign a tx

Move wallet test utility to wallet/ directory

Fix fails if the outpoint is invalid

which means that a tx the outpoint indicates is not exist in the wallet.